MDL-41803 lib: switch debugging() default to DEBUG_DEVELOPER
This function was frequently incorrectly used without DEBUG_DEVELOPER
for messages only developers could fix, so the default has switched.

As part of this change we've simply switch to an explicit DEBUG_NORMAL
to retain previous behaviour. Some of these uses would probably actually
better as a user level message. But we'll play it safe for now.