MDL-59648 core: Set autocomplete input size based on placeholder text
* Adjust the autocomplete form element's size depending on the
placeholder text's length.
* On the CSS side, set a minimum width of 200px for the autocomplete
form element. Set a maximum width of 90% of the container width as well
in order to prevent the down arrow from moving away from the input
element when the screen width is reduced. And if the width of the
autocomplete element gets narrower than the placeholder text,
apply ellipsis to the cutoff text.