MDL-51225 mod_forum: Wait before checking out recent activity block
To prevent random failures where the new forum post doesn't get shown
in the recent activity block, wait for 2 seconds before logging in as
the student and checking out the block.
Additionally, I edited the test so that $CFG->maxediting is directly set
to 1 second so the user would only have to wait for 2 seconds instead of
61 seconds in order to check the forum post editing time expiry. This
makes the test execution way faster.