MDL-71659 completion: Show debugging when grade item anomalies found
When there are grade item inconsistencies such as deleted grade items or
duplicated grade items. Errors will be thrown when  when fetching the
grade completion state for an activity. This breaks course and activity
views when an activity requires grade to determine its completion state.
So instead of throwing errors, show debugging instead.