MDL-71788 core: Create a copy to clipboard JS module
This works by passing the element IDs for the trigger and element that
contains the text to be copied.
A click event listener is attached to the trigger and checks for the
container element's value or innerText.
On success, a toast message is shown to the user that the text has
been copied to the clipboard.
When the element's value or innerText
cannot be found, a toast message indicating that the text cannot be
copied to the clipboard is shown.