MDL-76942 course: Always show the "add activity before" button
When editing is enabled on the course homepage, it is better to always
show the "Add an activity or resource before [xx]" buttons.
This is good for accessibility and usability as users can immediately
see that it is possible to add an activity or resource anywhere on the
course as opposed it being hidden where users may miss this feature
unless they accidentally hover or tab into the button, or unless they
know about the feature beforehand.