MDL-79056 core: Move combobox search result set ID to the container
The element ID for the combobox search result set is being used for the
aria-controls attribute of the search input. Currently, this ID is only
present when there are search results. If there are no search results,
the aria-controls attribute will be pointing to a non-existent element.
So it's better to point this to the results container.