MDL-79056 group: Render resultset initially
On initial render, the result set for the groups combobox search element
is not yet rendered. This causes the aria-controls attribute of the
search input element of the group combobox search to point to a non-
existent element and will break HTML validation. So render the result
set initially. The JS will update this container anyway when the
group combobox gets interacted on.