MDL-79061 course: Add sr-only ARIA live region for course announcements
When performing course actions, we can see the results of the action
visually. However, for visually impaired users, it is difficult for them
to know the results of the action, especially for AJAX actions that
cause changes on other parts of the page.

This patch adds a sr-only aria-live region so feedback from
AJAX actions can be conveyed to the screen reader users.