MDL-80231 ltix_gradebookservices: correct grade and grade item source
This value is arbitrary and ends up in the grade_grades_history and
grade_items_history tables. For grade_grades_history, this is visible,
as a string, in the grade history report. This change just makes sure
that it's using ltixservice_gradebookservices now, instead of
modltiservice_gradebookservices, and will not change existing history.