MDL-83470 composer: Require behat/gherkin 4.9.0 for now
`behat/gherkin` 4.10.0 introduces a breaking change where backslashes
in feature files need to be escaped. We need to pin the Behat version
to 4.9.0 for now and raise a separate issue to update to the later
versions once we've fixed the affected feature files and communicated
this change to the community.

More details:
https://github.com/Behat/Gherkin/releases/tag/v4.10.0